add new article page and clean up controller for articles

+++ app/controllers/articles_controller.rb
@@ -13,14 +13,18 @@ class ArticlesController < ApplicationController
     render json: { article: article.to_hash }
   end
 
+  def new
+    @article = Article.new
+  end
+
   def create
-    article = Article.new(article_params)
-    article.author_id = current_user.id
-    article.slug = article.generate_slug(article.title)
-    if article.save
-      render json: { article: article.to_hash }, status: :created
+    @article = Article.new(article_params)
+    @article.author_id = current_user.id
+    if @article.save
+      redirect_to article_path(@article.slug), notice: 'Article created successfully.'
     else
-      render json: { errors: article.errors.full_messages }, status: :unprocessable_entity
+      flash.now[:alert] = "There were errors saving your article."
+      render :new
     end
   end
 
@@ -34,8 +38,13 @@ class ArticlesController < ApplicationController
   end
 
   def destroy
-    article = current_user.articles.find_by_slug(params[:id])
-    article.destroy
+    article = Article.find_by_slug(params[:id])
+    if article && article.author_id == current_user.id
+      article.destroy
+    else
+      render json: { errors: ['Article not found'] }, status: :not_found
+      return
+    end
     head :no_content
   end
 
@@ -59,6 +68,6 @@ class ArticlesController < ApplicationController
   private
 
   def article_params
-    params.require(:article).permit(:title, :description, :body, tagList: [])
+    params.require(:article).permit(:title, :description, :body, tag_list: [])
   end
 end
